refactor: 모델 파일 통합작업
