+++ app/src/main/java/com/apptive/japkor/data/model/Matching.kt
@@ -1,5 +1,7 @@
 package com.apptive.japkor.data.model
 
+import com.google.gson.annotations.SerializedName
+
 data class MatchingResponse(
     val matchingId: Long,
     val maleMemberId: Long,
@@ -10,4 +12,20 @@ data class MatchingResponse(
     val residenceArea: String?,
     val matchingOrder: Int,
     val status: String
+)
+
+data class MalePendingMatchingResponse(
+    val matchingId: Long,
+    val femaleMemberId: Long,
+    val femaleName: String,
+    val femaleEmail: String,
+    val height: Int?,
+    val weight: Int?,
+    val residenceArea: String?,
+    val aiSummary: String?,
+    val status: String,
+    val createdAt: String,
+    val thumbnailImageUrl: String?,
+    @SerializedName(value = "profileImageUrls", alternate = ["profileImageUrl"])
+    val profileImageUrls: List<String>?
 )
