--- app/src/main/java/com/apptive/japkor/data/model/MyStatusResponse.kt
@@ -1,7 +1,0 @@
-package com.apptive.japkor.data.model
-
-data class MyStatusResponse(
-    val memberId: Long,
-    val name: String,
-    val status: UserStatus
-)
