+++ app/src/main/java/com/apptive/japkor/data/model/UserResponse.kt
@@ -1,5 +1,18 @@
 package com.apptive.japkor.data.model
 
+import com.google.gson.annotations.SerializedName
+
+data class AiSummaryResponse(
+    @SerializedName(value = "memberId", alternate = ["member_id"])
+    val memberId: Long,
+    @SerializedName(value = "name")
+    val name: String,
+    @SerializedName(value = "aiSummaryKo", alternate = ["ai_summary_ko", "aiSummary", "ai_summary"])
+    val aiSummaryKo: String?,
+    @SerializedName(value = "aiSummaryJa", alternate = ["ai_summary_ja"])
+    val aiSummaryJa: String?
+)
+
 data class MemberInfoResponse(
     val memberId: Long,
     val provider: String,
